Implemented CartProduct return on removeProduct from it's implementation

+++ src/main/java/com/gftworkshop/cartMicroservice/services/impl/CartProductServiceImpl.java
@@ -1,10 +1,13 @@
 package com.gftworkshop.cartMicroservice.services.impl;
 
+import com.gftworkshop.cartMicroservice.exceptions.CartProductNotFoundException;
 import com.gftworkshop.cartMicroservice.exceptions.CartProductSaveException;
 import com.gftworkshop.cartMicroservice.model.CartProduct;
 import com.gftworkshop.cartMicroservice.repositories.CartProductRepository;
 import com.gftworkshop.cartMicroservice.services.CartProductService;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 
 @Service
@@ -30,7 +33,20 @@ public class CartProductServiceImpl implements CartProductService {
     }
 
     @Override
-    public void removeProduct(Long id) {
-        cartProductRepository.deleteById(id);
+    public CartProduct removeProduct(Long id) {
+
+//        return cartProductRepository.deleteById(id);
+
+        Optional<CartProduct> optionalCartProduct = cartProductRepository.findById(id);
+
+        if (optionalCartProduct.isPresent()) {
+            CartProduct cartProduct = optionalCartProduct.get();
+
+            cartProductRepository.deleteById(id);
+
+            return cartProduct;
+        } else {
+            throw new CartProductNotFoundException("No se encontró el CartProduct con ID: " + id);
+        }
     }
 }
